~  Put all Actions related to game from the HomeController into a new Controller GameController, and moved and renamed Game View to /Game/Index ~ Added attribute CellCount to board and removed setters from Rows and Columns ~ Moved _Board one directory below ~ Readded site.css and .js to prevent errors
