+++ minesweeper/Program.cs
@@ -22,7 +22,7 @@ app.MapStaticAssets();
 
 app.MapControllerRoute(
 	name: "default",
-	pattern: "{controller=Home}/{action=Game}/{id?}")
+	pattern: "{controller=Home}/{action=Index}/{id?}")
 	.WithStaticAssets();
