+++ minesweeper/Views/Game/Index.cshtml
