+++ minesweeper/Views/Home/Index.cshtml
@@ -13,6 +13,7 @@
 }
 
 <div class="text-center">
-    <h1 class="display-4">Welcome</h1>
-    <p>Learn about <a href="https://learn.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
+    <h1 class="display-4">Welcome, fellow Sweeper!</h1>
+    <p>Click <a asp-controller="Game" asp-action="Index">here</a> to be directed to the Game page!</p>
+    <p>Or click <a asp-controller="Settings" asp-action="Index">here</a> to go to the Settings page!</p>
 </div>
